fix: 관리자, 회원 간에 로직이 동일한 api는 회원 쪽으로 옮기기
pr: #384
