+++ src/test/java/keeper/project/homepage/admin/controller/clerk/AdminSurveyControllerTest.java
@@ -257,85 +257,6 @@ public class AdminSurveyControllerTest extends SurveySpringTestHelper {
   }
 
   @Test
-  @DisplayName("[SUCCESS] 설문 정보 조회")
-  public void getSurveyInformation() throws Exception {
-    SurveyEntity survey = generateSurvey(LocalDateTime.now(), LocalDateTime.now().plusDays(2),
-        true);
-    generateSurveyMemberReply(survey, admin, surveyReplyRepository.getById(ACTIVITY.getId()));
-
-    mockMvc.perform(
-            get("/v1/admin/clerk/surveys/information/{surveyId}", survey.getId())
-                .header("Authorization", adminToken)
-                .contentType(MediaType.APPLICATION_JSON))
-        .andDo(print())
-        .andExpect(status().isOk())
-        .andExpect(jsonPath("$.success").value(true))
-        .andExpect(jsonPath("$.code").value(0))
-        .andDo(document("survey-admin-information",
-            pathParameters(
-                parameterWithName("surveyId").description("조회하는 설문 ID")
-            ),
-            responseFields(
-                generateSurveyInformationDtoResponseFields(ResponseType.SINGLE,
-                    "성공: true +\n실패: false",
-                    "성공 시 0을 반환",
-                    "성공: 성공하였습니다 +\n실패: 에러 메세지 반환")
-            )));
-  }
-
-  @Test
-  @DisplayName("[SUCCESS] 공개 상태의 설문 조회(가장 최근) - 현재 진행중인")
-  public void getLatestVisibleSurveyId() throws Exception {
-    SurveyEntity survey1 = surveyRepository.getById(ACTIVITY.getId());
-    SurveyEntity survey2 = generateSurvey(LocalDateTime.now().minusDays(2),
-        LocalDateTime.now().plusDays(2),
-        true);
-
-    mockMvc.perform(
-            get("/v1/admin/clerk/surveys/visible/ongoing")
-                .header("Authorization", adminToken)
-                .contentType(MediaType.APPLICATION_JSON))
-        .andDo(print())
-        .andExpect(status().isOk())
-        .andExpect(jsonPath("$.success").value(true))
-        .andExpect(jsonPath("$.code").value(0))
-        .andExpect(jsonPath("$.data").isNumber())
-        .andDo(document("survey-admin-visible-latest",
-            responseFields(
-                fieldWithPath("success").description("성공: true +\n실패: false"),
-                fieldWithPath("code").description("성공 시 0을 반환"),
-                fieldWithPath("msg").description("성공: 성공하였습니다 +\n실패: 에러 메세지 반환"),
-                fieldWithPath("data").description("조회한 설문 ID")
-            )));
-  }
-
-  @Test
-  @DisplayName("[SUCCESS] 종료된 설문 정보 조회(가장 최근)")
-  public void getLatestClosedSurveyInformation() throws Exception {
-    SurveyEntity survey1 = surveyRepository.getById(ACTIVITY.getId());
-    SurveyEntity survey2 = generateSurvey(LocalDateTime.now().minusDays(4),
-        LocalDateTime.now().minusDays(2),
-        true);
-    generateSurveyMemberReply(survey2, admin, surveyReplyRepository.getById(ACTIVITY.getId()));
-
-    mockMvc.perform(
-            get("/v1/admin/clerk/surveys/visible/closed")
-                .header("Authorization", adminToken)
-                .contentType(MediaType.APPLICATION_JSON))
-        .andDo(print())
-        .andExpect(status().isOk())
-        .andExpect(jsonPath("$.success").value(true))
-        .andExpect(jsonPath("$.code").value(0))
-        .andDo(document("survey-admin-closed-latest",
-            responseFields(
-                generateCloseSurveyInformationDtoResponseFields(ResponseType.SINGLE,
-                    "성공: true +\n실패: false",
-                    "성공 시 0을 반환",
-                    "성공: 성공하였습니다 +\n실패: 에러 메세지 반환")
-            )));
-  }
-
-  @Test
   @DisplayName("[SUCCESS] 설문 목록 조회")
   public void getSurveyList() throws Exception {
     Boolean isVisible = true;
